ClientBuilder moved inside Guzzle namespace
