+++ tests/Guzzle/fixtures/api_parts_manufacturers.json
