+++ tests/Route/fixtures/countries.json
@@ -1,0 +1,12 @@
+[
+  {
+    "id": "+39",
+    "code": "IT",
+    "name": "Italy"
+  },
+  {
+    "id": "+49",
+    "code": "DE",
+    "name": "Germany"
+  }
+]
